[CoreIPC] Messages::WebPageProxy::ShowContextMenuFromFrame allows arbitrary PDF rendering in UI process from web content https://bugs.webkit.org/show_bug.cgi?id=282103 rdar://135720091
Reviewed by Wenson Hsieh.

WebKit supplies images to the sharing service picker so that they can be shared
out to the system when the user right clicks on an image and clicks "Share".

Since the sharing service picker must be created in order to obtain the menu item
itself, the supplied `NSImage` is created when the context menu is displayed, rather
than when the user clicks "Share". This means that the code which creates the
`NSImage` is reachable simply with the `ShowContextMenuFromFrame` IPC message.
Additionally, the image is created with raw data rather than using a ShareableBitmap,
as the original format should be preserved when sharing. Consequently, the IPC
message can act as a way to perform image processing in the UI process.

To fix, supply a placeholder `NSImage` when creating the sharing service picker,
solely for the purpose of obtaining the menu item. Then, when the menu item is
clicked, another item is created with the real image, and then programmatically
invoked to present the sharing picker. This solution ensures that image processing
does not occur until the user actually clicks "Share".

* Source/WebKit/UIProcess/mac/WebContextMenuProxyMac.h:
* Source/WebKit/UIProcess/mac/WebContextMenuProxyMac.mm:
(-[WKMenuTarget performShare:]):
(WebKit::WebContextMenuProxyMac::handleShareMenuItem):

Set a `menu` on the created menu item, to ensure that the sharing service
picker has knowledge of the presenting view and the associated context menu.

The picker uses that information to make decisions about its presentation.

(WebKit::WebContextMenuProxyMac::createShareMenuItem):

Specify a placeholder image when needed, and simply grab the title of the sharing
menu item to create an item for display in the context menu.

(WebKit::WebContextMenuProxyMac::getContextMenuItem):
(WebKit::WebContextMenuProxyMac::getShareMenuItem): Deleted.
* Tools/TestWebKitAPI/Tests/mac/ContextMenuTests.mm:
(-[NSSharingServicePicker swizzled_initWithItems:]):
(TestWebKitAPI::TEST(ContextMenuTests, ShowSharePopoverForImage)):

Add a test to ensure that the sharing picker is still presented when right clicking
on an image and clicking share, and that the image is preserved.

Canonical link: https://commits.webkit.org/283286.355@safari-7620-branch