[watchOS] Occasional crash in ViewGestureGeometryCollector::computeTextLegibilityScales when double-tapping https://bugs.webkit.org/show_bug.cgi?id=288407 rdar://145501290
Reviewed by Abrar Rahman Protyasha.

Avoid hash table lookup and mutation in `computeTextLegibilityScales`, in the case where the font
size is not a valid `unsigned` hash key; this would correspond to either 0 or an extremely large
value anyways, which can safely be ignored for the purposes of this heuristic, which attempts to
estimate a good scale to set when double-tapping to zoom that makes most of the text on the webpage
legible, based on a histogram of sampled font sizes for visible text in the DOM.

To make this code a bit easier to reason about, we also replace the `HashMap<unsigned, unsigned>`
with `HashCountedSet<unsigned>`, and use `add(…)` to increment the counter per bucket. Importantly,
we can also use `HashCountedSet::isValidValue` to avoid any attempts to perform hash table lookup/
mutations against invalid keys (note that the `ValueType` for a `HashCountedSet` corresponds to the
hash key, not the internal count which is always `unsigned`).

* Source/WebKit/WebProcess/WebPage/ViewGestureGeometryCollector.cpp:
(WebKit::ViewGestureGeometryCollector::computeTextLegibilityScales):

Canonical link: https://commits.webkit.org/289651.168@safari-7621-branch