[CoreIPC][Networking] MemoryObjectStoreCursor::iterate should invalidate `m_iterator` too when invalidating `m_currentPositionKey` https://bugs.webkit.org/show_bug.cgi?id=289111 rdar://143027372
Reviewed by Sihui Liu.

`m_currentPositionKey` must always be in sync with `m_iterator`. As in, it must always be
equal to the key that `m_iterator` is currently pointing to. It follows that when invalidating
`m_currentPositiionKey` in MemoryObjectStoreCursor::iterate(), we must invalidate `m_iterator`
as well.

* Source/WebCore/Modules/indexeddb/server/MemoryObjectStoreCursor.cpp:
(WebCore::IDBServer::MemoryObjectStoreCursor::iterate):

Canonical link: https://commits.webkit.org/289651.236@safari-7621-branch