WebGL: Generating mipmaps for 3D textures with depth greater than width, height causes crash https://bugs.webkit.org/show_bug.cgi?id=289677 rdar://145485379
Reviewed by Mike Wyrzykowski.

When 3D texture depth > max(2 * w, 2 * h), the last levels ended up
calling the mipmap generation function with sourceWidth=1,
sourceHeight=1, sourceDepth=1. This is unanticipated, as that mipmap
is the last mipmap level and cannot be a source for mipmap generation.
This leads to the mipmap generator to select a nullptr function
and calling it.

This happened in certain conditions depending on features, internal
format and texture size. The conditions would make the mipmap generation
take the CPU mipmap generation path.

* Source/ThirdParty/ANGLE/src/libANGLE/renderer/metal/TextureMtl.mm:
(rx::TextureMtl::generateMipmapCPU):
* Source/ThirdParty/ANGLE/src/tests/gl_tests/MipmapTest.cpp:

Canonical link: https://commits.webkit.org/289651.278@safari-7621-branch