Unreviewed. [WPE] Use Skia with CPU rendering by default
The WPE port has been found to perform better when using the CPU with
Skia for painting, instead of the default (GPU), in a number of typical
embedded devices. This ensures that we do not regress on performance due
to weak GPUs in the 2.48 release branch.

* Source/WebKit/WebProcess/glib/WebProcessGLib.cpp:
  (WebKit::WebProcess::platformInitializeWebProcess): Change the default
  to use CPU rendering in the WPE port, and handle the case in which
  WEBKIT_SKIA_ENABLE_CPU_RENDERING=0 is set in the environment to switch
  to GPU rendering.

Canonical link: https://commits.webkit.org/290945.64@webkitglib/2.48