CrashTracer: com.apple.WebKit.Networking.Development at com.apple.WebKit: WTF::assertIsCurrent rdar://102953242
Reviewed by Chris Dumez.

Sometimes, the completion handler finalizer will be executed out of the main thread as the block is sometimes destroyed out of the main thread.
To cope with this, we update CompletionHandlerWithFinalizer to allow disallowing thread safety assertions.
And we change the finalizer to hop to main thread explicitly.

* Source/WTF/wtf/CompletionHandler.h:
(WTF::CompletionHandlerWithFinalizer<Out):
* Source/WebKit/NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::NetworkDataTaskCocoa::setH2PingCallback):

Canonical link: https://commits.webkit.org/289651.305@safari-7621-branch