Fix heap-uaf in computeCompositingRequirements https://bugs.webkit.org/show_bug.cgi?id=287908 <rdar://144404223>
Reviewed by Matt Woodrow.

This issue was caused by us clearing the backing store of a layers
positive z-order lists while recursing through the layer tree. More
specifically, we need to make sure to dirty all ancestor layers when
content visibility changes, before we start traversing the tree, and
this was not being done for ancestor layers in
`setIsHiddenByOverflowTruncation`.

Change this function to let all ancestor layers know that they have a
descendant with a dirty content visibility status, which will allow us
to get the tree into the proper state when calling
`updateDescendantDependentFlags` on the root layer and having no dirty
flags in the tree following this.

* Source/WebCore/rendering/RenderLayer.h:
(WebCore::RenderLayer::setIsHiddenByOverflowTruncation):

Canonical link: https://commits.webkit.org/289651.364@safari-7621-branch