WKFrameInfo.request should never return nil https://bugs.webkit.org/show_bug.cgi?id=291334 rdar://148793218
Reviewed by Wenson Hsieh.

The attribute is annotated as non-nullable, which means if Swift tries to access
it it won't check null first.  If we return null, it will crash.

rdar://144600565 made some needed and correct changes to what frame info is returned
in some cases, and it also made it so that WKNavigationAction.sourceFrame is non-null
in more cases, which should also fix some Swift crashes.  However, it introduced
some cases where we have a non-null WKFrameInfo with a null request, which introduced
some crashes.  This should fix those crashes by returning about:blank for the request
that loaded a frame before it shows anything.

* Source/WebKit/UIProcess/API/Cocoa/WKFrameInfo.mm:
(-[WKFrameInfo request]):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/DecidePolicyForNavigationAction.mm:
(TEST(WebKit, NavigationActionFrames)):

Canonical link: https://commits.webkit.org/289651.419@safari-7621-branch