Unreviewed, fix build break after 292734@main rdar://148941035
In 286164@main, the calls to lambda isHDRColorSpace() was replaced by calling a
new function called DestinationColorSpace::usesStandardRange(). This replacement
for cleaning up the code and delete repeated calls to CG APIs.

But this new function was deleted in 292734@main. This caused a compilation error
on macOS Ventura. So restore the check for HDR in NativeImage::draw() as it was
before 286164@main.

* Source/WebCore/platform/graphics/cg/NativeImageCG.cpp:
(WebCore::NativeImage::draw):

Canonical link: https://commits.webkit.org/289651.420@safari-7621-branch